Rename source file and add reuse stuff
