+++ src/pypdfium2/_helpers/text_inserter.py
